Fix rust kernel build and add rust-analyzer support
Kernel build now generates the config correctly. It works by disabling
GCC_PLUGINS which allows for it to be configured through `make config`.
A patch is also applied to the bindgen libclang version checker to fix
a bug that appears only in `dev` outputs. In dev outputs there is a
version in the path which gets picked up instead of the actual version.
This fixes the dev outputs and no longer requires manually copying build
files from the original /build directory.